test(integration): load C81183 monitor before Alloy
